change get api deal with the 500 code
If you cat a wrong hash, the ipfs http api will response code 500
and with message. In this case if we use the response boy and cover it
to stream or byte ,it will not the true content of the hash code,it will
be the content of the 'Wrong Message'.

So we need check the response code by {response.isSuccessful} to insure
we go the right response.